pyproject.toml
[project]
name = "project2-resume-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openai>=1.106.1",
    "pymupdf>=1.26.4",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "tenacity>=9.1.2",
    "truststore>=0.10.4",
]
